Updated MenuActivity .
