+++ app/src/main/res/layout/card_view_row.xml
@@ -4,13 +4,13 @@
     android:orientation="vertical"
 
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="wrap_content">
 
     <android.support.v7.widget.CardView
         android:id="@+id/card_view"
         android:layout_gravity="center"
         android:layout_width="fill_parent"
-        android:layout_height="150dp"
+        android:layout_height="149dp"
         android:layout_margin="5dp"
         android:elevation="5dp"
         card_view:cardCornerRadius="2dp"
@@ -18,7 +18,7 @@
 
         <RelativeLayout
             android:layout_width="fill_parent"
-            android:layout_height="fill_parent">
+            android:layout_height="wrap_content">
 
             <TextView
                 android:id="@+id/textView"
