+++ app/src/main/res/values/strings.xml
@@ -40,5 +40,6 @@
     <string name="navigation_drawer_close">Close navigation drawer</string>
 
     <string name="action_settings">Settings</string>
+    <string name="title_activity_menu">MenuActivity</string>
 
 </resources>
